requirements.txt
pytest>=8.0
requests>=2.31
python-dotenv>=1.0
jsonschema>=4.21
allure-pytest>=2.13
pytest-playwright>=0.5
